# Optimize Migration Script