Benchmarks
===========
For a staging dump with ~60000 `NodeLogs`:
```
1) No optimization, just error handling:
 960.45s user 112.76s system 86% cpu 20:37.76 total
~4.5 hr on prod

2) Logging once per 1000 original nodelogs:
746.36s user 79.38s system 88% cpu 15:31.23 total
~3.4 hr on prod

3) No logging whatsoever:
737.77s user 77.76s system 89% cpu 15:12.22 total
~3.3 hr on prod

4) No logging, using pymongo for updates / inserts:
23.99s user 3.35s system 41% cpu 1:05.11 total
~6.66 min on prod

5) Logging all _id's and errors, using pymongo for updates / inserts:
71.48s user 8.50s system 52% cpu 2:31.12 total
~19.86 min on prod
```

Updates
=======
 # 5 implemented.
Manually tested to verify pre-existing logs were working after this script was run.